write a new getFriendByName on friendDao,write native(c) lib and java(mainActivity) to add full debt(date,amount,etc.) all at once , do some test(pass) ,AppDatabase.java write a tempoary line to wipe all db for testing

+++ app/src/main/java/com/example/dept_app/data/Debts.java
@@ -21,7 +21,7 @@ public class Debts {
     private int friendsId;
     private String date;
     private String description;
-    private int amount;
+    private double amount;
     private String type; // "you_owe" or "friend_owes"
 
     // Getters and setters
@@ -37,8 +37,8 @@ public class Debts {
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
 
-    public int getAmount() { return amount; }
-    public void setAmount(int amount) { this.amount = amount; }
+    public double getAmount() { return amount; }
+    public void setAmount(double amount) { this.amount = amount; }
 
     public String getType() { return type; }
     public void setType(String type) { this.type = type; }
